Changed cast type

+++ template_based_SFT/finetune_peft_classification.py
@@ -120,10 +120,10 @@ def main():
 
     if "gpt-neox" in model_args.model_name_or_path:
         model = prepare_model_for_int8_training(
-            model, output_embedding_layer_name="embed_out", layer_norm_names=["layer_norm", "layernorm"], cast_dtype=torch.float16
+            model, output_embedding_layer_name="embed_out", layer_norm_names=["layer_norm", "layernorm"], cast_dtype=torch.float32
         )
     else:
-        model = prepare_model_for_int8_training(model, cast_dtype=torch.float16)
+        model = prepare_model_for_int8_training(model, cast_dtype=torch.float32)
 
 
     # ### Apply LoRA
